Separete config file
