+++ configs/hsts.js
@@ -1,0 +1,5 @@
+module.exports = {
+  maxAge: 10886400000,
+  includeSubDomains: true,
+  force: true
+};
