+++ configs/meta.js
@@ -1,0 +1,13 @@
+module.exports = {
+  name: 'mdl-skeleton',
+  alias: '',
+  slogan: '',
+  description: '',
+  keywords: '',
+  image: '',
+  author: '',
+  title: 'Material Design Lite',
+  domain: '',
+  url: '',
+  canonical: ''
+};
