+++ configs/settings.js
@@ -1,0 +1,5 @@
+module.exports = {
+  dateformat: 'D, M d, Y h:i:s O',
+  viewDir: './app/views/',
+  tplFileExtension: '.html'
+};
