+++ configs/static.js
@@ -1,0 +1,5 @@
+module.exports = {
+  maxAge: 24 * 60 * 6e4,
+  etag: true,
+  lastModified: true
+};
