--- configs/vars.sample.js
@@ -1,11 +1,0 @@
-var domain = 'md.techpush.me';
-
-module.exports = {
-  ENV: 'dev',
-  port: 8889,
-  meta: {
-    domain,
-    image: 'http://' + domain + '/images/bg/bb.jpg',
-    url: 'http://' + domain
-  }
-};
